Move control into show
