+++ app/Http/Livewire/Printers/Show.php
@@ -3,11 +3,15 @@
 namespace App\Http\Livewire\Printers;
 
 use App\Models\Printer;
+use Filament\Notifications\Notification;
+use Illuminate\Support\Facades\Http;
 use Livewire\Component;
 
 class Show extends Component
 {
     public Printer $printer;
+
+    public $amount = 10;
 
     public function render()
     {
@@ -26,5 +30,51 @@ class Show extends Component
         $this->printer->safeDelete();
 
         return redirect('printers');
+    }
+
+    public function home($axis)
+    {
+        $response = Http::octoPrint($this->printer)->post('api/printer/printhead', [
+            'command' => 'home',
+            'axes' => $axis,
+        ]);
+
+        if ($response->failed()) {
+            Notification::make()
+                ->title($response->json('error'))
+                ->danger()
+                ->send();
+        }
+    }
+
+    public function move($axis, $direction = '')
+    {
+        $value = $direction . $this->amount;
+
+        if ($axis === 'x') {
+            $this->jog((float) $value, 0, 0);
+        } elseif ($axis === 'y') {
+            $this->jog(0, (float) $value, 0);
+        } elseif ($axis === 'z') {
+            $this->jog(0, 0, (float) $value);
+        }
+    }
+
+    private function jog($x, $y, $z)
+    {
+        $response = Http::octoPrint($this->printer)
+            ->post('api/printer/printhead', [
+                'command' => 'jog',
+                'x' => $x,
+                'y' => $y,
+                'z' => $z,
+            ]);
+
+        if ($response->failed()) {
+            Notification::make()
+                ->title($response->json('error'))
+                ->danger()
+                ->send();
+        }
     }
 }
