+++ resources/views/livewire/printers/partials/control.blade.php
