+++ resources/views/livewire/printers/show.blade.php
@@ -17,7 +17,7 @@
 
         <div class="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-4">
             @include('livewire.printers.partials.webcam')
-            <livewire:printers.control :printer="$printer" />
+            @include('livewire.printers.partials.control')
             <livewire:printers.temps :printer="$printer" />
             <livewire:bit.assign-material :tools="$tools" />
             <livewire:printers.connection :printer="$printer" />
